requirements.txt
tortoise-orm==0.17.0
fastapi
pydantic
tortoise-orm[asyncpg]
requests
redis
uvicorn
